updated implementation to match #149
